+++ NegocioTutorias/src/main/java/Interfaces/ITutorBO.java
@@ -16,6 +16,6 @@ public interface ITutorBO {
     public TutorDTO registrarTutor(TutorDTO tutor) throws NegocioException;
     public ArrayList<TutorDTO> consultarTodos() throws NegocioException;
     public TutorDTO modificarTutor(TutorDTO tutor) throws NegocioException;
-    public TutorDTO eliminarTutor(long id) throws NegocioException;
+    public TutorDTO eliminarTutor(int id) throws NegocioException;
     public TutorDTO buscarPorId(long id) throws NegocioException;
 }
